service: Deallocate database on separate thread
For databases that we're done with, we can safely create a throwaway coroutine for deallocating the database. This can take a while for very large databases (500k entries or more).

+++ service/src/main/kotlin/dev/clombardo/dnsnet/service/db/RuleDatabaseManager.kt
@@ -89,10 +89,16 @@ class RuleDatabaseManager(
     fun setShouldStop(shouldStop: Boolean) = ruleDatabaseController.setShouldStop(shouldStop)
 
     fun destroy() {
-        onLoadedFiltersChanged(0UL)
+        if (destroyed) {
+            return
+        }
         destroyed = true
+        setShouldStop(true)
         waitOnInit()
-        ruleDatabase.destroy()
-        ruleDatabaseController.destroy()
+        onLoadedFiltersChanged(0UL)
+        CoroutineScope(Dispatchers.IO).launch {
+            ruleDatabase.destroy()
+            ruleDatabaseController.destroy()
+        }
     }
 }
